Move `NativeAppStartEventProcessor` functionality into integration
