+++ flutter/test/sentry_navigator_observer_test.dart
@@ -7,7 +7,6 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
-import 'package:sentry_flutter/src/integrations/integrations.dart';
 import 'package:sentry/src/sentry_tracer.dart';
 import 'package:sentry_flutter/src/native/native_frames.dart';
 import 'package:sentry_flutter/src/navigation/time_to_display_tracker.dart';
@@ -489,16 +488,6 @@ void main() {
 
     test('flutter root name is replaced', () async {
       final rootRoute = route(RouteSettings(name: '/'));
-      NativeAppStartIntegration.setAppStartInfo(
-        AppStartInfo(
-          AppStartType.cold,
-          start: DateTime.now().add(const Duration(seconds: 1)),
-          end: DateTime.now().add(const Duration(seconds: 2)),
-          pluginRegistration: DateTime.now().add(const Duration(seconds: 3)),
-          sentrySetupStart: DateTime.now().add(const Duration(seconds: 4)),
-          nativeSpanTimes: [],
-        ),
-      );
 
       final hub = _MockHub();
       final span = getMockSentryTracer(name: '/');
